requirements.txt
# 核心依赖
crawl4ai>=0.6.1
openai>=1.0.0
python-dotenv>=1.0.0

# 网络请求
requests>=2.31.0
aiohttp>=3.9.0

# 内容解析
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 工具库
asyncio>=3.4.3
typing-extensions>=4.8.0
